chore: more file restructuring and organization
