+++ lib/utils/xcpUtils.ts
@@ -1,6 +1,10 @@
-import { xcp_public_nodes, xcp_v2_nodes } from "$server/services/xcpService.ts";
+// lib/utils/xcpUtils.ts
+
+// TODO: move to server for xcp related functions
+
+import { xcp_public_nodes } from "$server/services/xcpService.ts";
 import { XCPParams } from "globals";
-import { dbManager } from "$server/db.ts";
+import { dbManager } from "$server/database/index.ts";
 
 export interface XCPPayload {
   jsonrpc: string;
@@ -9,12 +13,6 @@ export interface XCPPayload {
   params: XCPParams;
 }
 
-/**
- * Creates a payload object for JSON-RPC requests.
- * @param method - The method name.
- * @param params - The parameters for the method.
- * @returns The payload object.
- */
 export const CreatePayload = (
   method: string,
   params: XCPParams,
@@ -44,14 +42,6 @@ export async function handleXcpApiRequestWithCache(
   );
 }
 
-/**
- * Makes a query to the specified URL with retries in case of failure.
- * @param url - The URL to make the query to.
- * @param auth - The authentication string.
- * @param payload - The payload to send with the query.
- * @param retries - The number of retries to attempt (default is 0).
- * @returns The result of the query or null if all retries failed.
- */
 const handleXcpQueryWIthRetries = async (
   url: string,
   auth: string,
@@ -91,12 +81,6 @@ const handleXcpQueryWIthRetries = async (
   }
 };
 
-/**
- * Handles the query by sending it to multiple public nodes and returning the result from the first successful query.
- * If all queries fail, it logs an error message and returns null.
- * @param payload - The query payload to be sent to the nodes.
- * @returns The result of the successful query or null if all queries fail.
- */
 export const handleXcpV1Query = async (payload: XCPPayload): Promise<any> => {
   console.log("Starting handleXcpV1Query with payload:", payload);
 
@@ -111,7 +95,6 @@ export const handleXcpV1Query = async (payload: XCPPayload): Promise<any> => {
         0,
       );
 
-      // If we get a result (even if it's an error), return it
       if (result !== null) {
         console.log(`Response from node ${node.url}:`, result);
         return result;
@@ -119,16 +102,13 @@ export const handleXcpV1Query = async (payload: XCPPayload): Promise<any> => {
     } catch (error) {
       console.error(`Error querying node ${node.url}:`, error);
 
-      // If it's not a connection error, throw the error
       if (
         !(error instanceof TypeError && error.message.includes("connection"))
       ) {
         throw error;
       }
-      // If it is a connection error, we'll try the next node
     }
   }
 
-  // If we've tried all nodes and none worked
   throw new Error("All queries to nodes have failed due to connection issues.");
 };
