+++ server/database/index.ts
@@ -1,5 +1,8 @@
+export { dbManager } from "./databaseManager.ts";
 export { StampRepository } from "./stampRepository.ts";
 export { BlockRepository } from "./blockRepository.ts";
 export { CollectionRepository } from "./collectionRepository.ts";
-
 export { summarize_issuances } from "./summary.ts";
+
+// Export types if needed
+export type { DatabaseConfig } from "./databaseManager.ts";
