+++ server/services/src20/psbt/index.ts
@@ -1,0 +1,3 @@
+import { SRC20PSBTCreateService } from "./createService.ts";
+
+export { SRC20PSBTCreateService };
