+++ server/services/transaction/index.ts
@@ -1,0 +1,9 @@
+import { UTXOService } from "./utxoService.ts";
+import { PSBTService } from "./psbtService.ts";
+
+export { UTXOService, PSBTService };
+
+export class TransactionService {
+  static readonly UTXOService = UTXOService;
+  static readonly PSBTService = PSBTService;
+}
